+++ docs/history.md
@@ -1,3 +1,7 @@
+## 1.2.4
+* Added a arrowTrayectoryPoints in the result of getMasterGrade function
+* Added example of arrow trayecotry preview on bow.js example
+
 ## 1.2.3
 * Fixed bug with crossbow enchant